pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "branchwise"
version = "0.1.0"
description = "An intelligent code review agent."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Or", email = "or@example.com"}
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyGithub>=2.1.1",
    "python-gitlab>=3.0.0",
    "atlassian-python-api>=3.0.0",
    "openai>=1.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0"
]

[project.scripts]
branchwise = "branchwise.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black",
    "isort",
    "mypy"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["branchwise*"]
